Updated coin appearance and behavior

+++ src/tmp/GameObject.java
@@ -32,10 +32,10 @@ public abstract class GameObject {
 		public abstract Polygon getBounds();
 
 		//Get/Set
-		public void setX(int x) {
+		public void setX(float x) {
 			this.x = x;
 		}
-		public void setY(int y) {
+		public void setY(float y) {
 			this.y = y;
 		}
 		public void setWidth(int width) { this.width = width; }
@@ -44,10 +44,10 @@ public abstract class GameObject {
 		public void setID(ID id) {
 			this.id = id;
 		}
-		public void setVelX(int velX) {
+		public void setVelX(float velX) {
 			this.velX = velX;
 		}
-		public void setVelY(int velY) {
+		public void setVelY(float velY) {
 			this.velY = velY;
 		}
 		public void setxDirection(int xDirection) {
